Phase_3_Data_Warehouse_and_Analytics_Tools/requirements.txt
# AWS SDK for Python
boto3>=1.26.0
botocore>=1.29.0

# Snowflake connectivity
snowflake-connector-python>=3.0.0

# PostgreSQL connectivity
psycopg2-binary>=2.9.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Additional utilities
python-dotenv>=1.0.0 